requirements.txt
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
psutil==5.9.6
pyautogui==0.9.54
python-dotenv==1.0.0 